feat: move over to single file for all tests types
